+++ simple_manga_downloader/modules/__init__.py
@@ -5,3 +5,4 @@ from .mangakakalot_com import Mangakakalot
 from .manganelo_com import Manganelo
 from .config_parser import Config
 from .manga import BaseManga
+from .argparser import parse_arguments