+++ simple_manga_downloader/modules/__init__.py
@@ -3,6 +3,4 @@ from .mangaseeonline_us import Mangasee
 from .mangatown_com import Mangatown
 from .mangakakalot_com import Mangakakalot
 from .manganelo_com import Manganelo
-from .config_parser import Config
 from .manga import BaseManga
-from .argparser import parse_arguments